Move entry component into folder
