+++ passthepassword-frontend/src/components/entry/Entry.tsx
