+++ passthepassword-frontend/src/components/entry/entry.type.tsx
